API: change types of "start" and "length" arguments for "setInstrumentLoopData" from int to int64_t

+++ audio/Sampler_plugin_proc.h
@@ -23,7 +23,7 @@ extern LANGSPEC void SAMPLER_erase_recorded_peaks(SoundPlugin *plugin);
 
 extern LANGSPEC bool SAMPLER_set_new_sample(struct SoundPlugin *plugin, const wchar_t *filename, int instrument_number);
 extern LANGSPEC bool SAMPLER_set_random_sample(struct SoundPlugin *plugin, const wchar_t *path);
-extern LANGSPEC void SAMPLER_set_loop_data(struct SoundPlugin *plugin, int start, int length);
+extern LANGSPEC void SAMPLER_set_loop_data(struct SoundPlugin *plugin, int64_t start, int64_t length);
 
 extern LANGSPEC bool SAMPLER_set_temp_resampler_type(struct SoundPlugin *plugin, enum ResamplerType resampler_type);
 extern LANGSPEC void SAMPLER_set_org_resampler_type(struct SoundPlugin *plugin);
